Added MongoCommandException CommandFailureException extends MongoCommandException and is deprecated UnacknowledgedWriteException extends UnsupportedOperationException instead of MongoClientException and is deprecated
  JAVA-1570
