+++ src/main/com/mongodb/MongoClientException.java
@@ -18,6 +18,8 @@ package com.mongodb;
 
 /**
  * A base class for exceptions indicating a failure condition within the driver.
+ *
+ * @since 2.12
  */
 public class MongoClientException extends MongoInternalException {
